+++ tests/test_mara_markdown_docs.py
@@ -28,6 +28,7 @@ def _client():
             'xyz_Marketing_xyz': repo_root_dir / 'app/pipelines/marketing/README.md',
             'Developer/Setup': repo_readme,
             'Developer/Code Conventions': repo_root_dir / 'code_conventions.md',
+            'test': repo_root_dir / 'tests/docs/test.md'
         }
 
         return docs
@@ -40,12 +41,20 @@ def _client():
 
 client = pytest.fixture(_client)
 
+def test_redirect_to_full_url(client):
+    r = client.get('/docs/developer')
+    print(r.data)
+    assert b'/docs/developer/' in r.data, 'no redirect'
+    assert b'Redirect' in r.data, 'no redirect'
+    assert r.status_code == 308, 'No redirect: status'
+
+
 def test_docs_in_menu(client):
     r = client.get('/mara-app/navigation-bar')
     print(r.data)
     assert b'>Developer</' in r.data, 'Missing folder structure'
     assert b'xyz_Marketing_xyz' in r.data, 'missing entry'
-    assert b'href="/docs/developer/code_conventions"' in r.data, 'missing URL'
+    assert b'href="/docs/developer/code_conventions/"' in r.data, 'missing URL'
     assert b'href="/mara-app/configuration#mara_markdown_docs.config"' in r.data, 'missing entry for config'
 
 
@@ -56,7 +65,7 @@ def test_docs_in_config(client):
     assert str(repo_readme).encode('utf8') in r.data, 'missing repo readme entry'
 
 def test_doc(client):
-    r = client.get('/docs/developer/setup')
+    r = client.get('/docs/developer/setup/')
     print(r.data)
     assert b'mara-markdown-docs.git' in r.data, 'Missing link to repo'
 
@@ -64,20 +73,50 @@ def test_overview(client):
     r = client.get('/docs/')
     print(r.data)
     assert b'xyz_Marketing_xyz' in r.data, 'missing entry'
-    assert b'href="/docs/developer/code_conventions"' in r.data, 'missing URL'
+    assert b'href="/docs/developer/code_conventions/"' in r.data, 'missing URL'
+
 
 def test_404_directory(client):
-    r = client.get('/docs/developer')
+    r = client.get('/docs/developer/')
     print(r.data)
     assert b'404' in r.data, 'Missing page is not missing'
     assert r.status_code == 404, 'Missing page is not missing: status'
+
 
 def test_404_page(client):
-    r = client.get('/docs/developer/missing')
+    r = client.get('/docs/developer/missing/')
     print(r.data)
     assert b'404' in r.data, 'Missing page is not missing'
     assert r.status_code == 404, 'Missing page is not missing: status'
 
+def test_docs_file_endpoint(client):
+    r = client.get('/docs/test/test.txt')
+    print(r.data)
+    assert b'$Test doc$' in r.data
+    assert r.status_code == 200
+
+def test_docs_file_endpoint_404(client):
+    r = client.get('/docs/test/test2.txt')
+    print(r.data)
+    assert r.status_code == 404
+
+def test_docs_file_endpoint_dot_dot(client):
+    # flask prevents / in the last segment, so this also guards against ..
+    r = client.get('/docs/test/../docs/test.txt')
+    print(r.data)
+    assert r.status_code == 404
+
+def test_docs_file_endpoint_subfolder(client):
+    # flask prevents / in the last segment, so also no subfolder
+    r = client.get('/docs/test/docs/test.txt')
+    print(r.data)
+    assert r.status_code == 404
+
+def test_docs_file_endpoint_not_allowed_filetype(client):
+    r = client.get('/docs/test/secrets.conf')
+    print(r.data)
+    assert b'$secrets$' not in r.data
+    assert r.status_code == 405
 
 @pytest.fixture
 def client_unauth():