+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## unreleased
+* Serve images and `.txt` files from the same folder as the markdown document
+
 ## 0.1.0 (2020-08-22)
 
 - Move to Github